ingestion/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.34
httpx>=0.27
asyncpg>=0.30
qdrant-client>=1.12
presidio-analyzer>=2.2
presidio-anonymizer>=2.2
spacy>=3.7
pydantic>=2.0
pyyaml>=6.0
pyjwt[crypto]>=2.8
cachetools>=5.3
prometheus-client>=0.21
opentelemetry-api>=1.20
opentelemetry-sdk>=1.20
opentelemetry-exporter-otlp-proto-grpc>=1.20
opentelemetry-instrumentation-fastapi>=0.41b0
opentelemetry-instrumentation-httpx>=0.48b0
